튜토리얼 완성 플레이어 hit, Ex공격 완성 world맵 플레이어 애니메이션 이동 완성 slimeBoss 맵 완성

+++ Assets/Script/Interface/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void OnDamage(int damage, Vector2 hitPoint);
+    void OnDamage(float damage, Vector2 hitPoint);
 }
